Convert RazorEditorFactoryService to interface
