+++ src/Razor/src/Microsoft.AspNetCore.Razor.ExternalAccess.LegacyEditor/RazorServices.cs
@@ -24,7 +24,7 @@ internal sealed class RazorServices
         var componentModel = serviceProvider.GetService(typeof(SComponentModel)) as IComponentModel;
         Assumes.Present(componentModel);
 
-        var editorFactoryService = componentModel.GetService<RazorEditorFactoryService>();
+        var editorFactoryService = componentModel.GetService<VisualStudio.Editor.Razor.IRazorEditorFactoryService>();
         EditorFactoryService = RazorWrapperFactory.WrapEditorFactoryService(editorFactoryService);
 
         var clientSettingsManager = componentModel.GetService<IClientSettingsManager>();
