+++ src/Razor/src/Microsoft.AspNetCore.Razor.ExternalAccess.LegacyEditor/RazorWrapperFactory.EditorFactoryServiceWrapper.cs
@@ -2,14 +2,13 @@
 // Licensed under the MIT license. See License.txt in the project root for license information.
 
 using System.Diagnostics.CodeAnalysis;
-using Microsoft.VisualStudio.Editor.Razor;
 using Microsoft.VisualStudio.Text;
 
 namespace Microsoft.AspNetCore.Razor.ExternalAccess.LegacyEditor;
 
 internal static partial class RazorWrapperFactory
 {
-    private class EditorFactoryServiceWrapper(RazorEditorFactoryService obj) : Wrapper<RazorEditorFactoryService>(obj), IRazorEditorFactoryService
+    private class EditorFactoryServiceWrapper(VisualStudio.Editor.Razor.IRazorEditorFactoryService obj) : Wrapper<VisualStudio.Editor.Razor.IRazorEditorFactoryService>(obj), IRazorEditorFactoryService
     {
         public bool TryGetDocumentTracker(ITextBuffer textBuffer, [NotNullWhen(true)] out IRazorDocumentTracker? documentTracker)
         {
